refactor: moved out services, clean up
